Shifted Mongo functions into DB lib and standardsised on OPc
