--- mongodb.yaml
@@ -1,14 +1,0 @@
-### MongoDB Backend ###
-##PyHSS has support for MongoDB backend instead of just the CSV file. This is useful when dealing with larger databases and for scaling.
-##PyHSS uses the same MongoDB data format as Open5Gs, so the Open5Gs Web Interface can be used to change the data.
-##There is also a Python library which includes info on adding / editing / deleting subscribers available here: https://github.com/open5gs/open5gs/tree/master/misc/mongodb/python
-##
-### Setting up your server to use MongoDB:
-## Install MongoDB (ie apt-get install mongodb)
-## Bind MongoDB to listen on a socket (edit /etc/mongodb.conf & set 'bind_ip = 127.0.0.1' and 'port = 27017' & restart MongoDB)
-## Install Python requirements with Pip (pip3 install pyyaml mongo)
-## Add some subscribers to the database using Open5Gs Web UI or Open5Gs Python Library (See example in /lib directory)
-mongodb_server: 127.0.0.1
-mongodb_username: root
-mongodb_password: password
-mongodb_port: 27017
